Simplify interfaces
