+++ kasa/tests/fakeprotocol_smart.py
@@ -218,7 +218,9 @@ class FakeSmartTransport(BaseTransport):
                 # SMART fixtures started to be generated
                 missing_result := self.FIXTURE_MISSING_MAP.get(method)
             ) and missing_result[0] in self.components:
-                result = copy.deepcopy(missing_result[1])
+                # Copy to info so it will work with update methods
+                info[method] = copy.deepcopy(missing_result[1])
+                result = copy.deepcopy(info[method])
                 retval = {"result": result, "error_code": 0}
             else:
                 # PARAMS error returned for KS240 when get_device_usage called
@@ -238,6 +240,10 @@ class FakeSmartTransport(BaseTransport):
             return {"error_code": 0}
         elif method == "set_dynamic_light_effect_rule_enable":
             self._set_light_effect(info, params)
+            return {"error_code": 0}
+        elif method == "set_led_info":
+            info["get_led_info"]["led_status"] = params["led_rule"] != "never"
+            info["get_led_info"]["led_rule"] = params["led_rule"]
             return {"error_code": 0}
         elif method[:4] == "set_":
             target_method = f"get_{method[4:]}"
